sparse_mmap: remove C code and add SAFETY comments
To a certain extent this change "hardcodes"/"emulates"
`setjmp` and `longjmp` (not their `sig` counterparts though).
These might be tightly intergrated with the system C library
making this code brittle. Need to look more into that.

The unit-test do pass on Linux and macOS, but sometimes the
Windows guest wouldn't boot in the integration tests (tested
only in nested so far). I might be cargo-culting something.