RFC: enable openhcl to run in VTL0, use nested virtualization
This change enables OpenHCL to be launched in a non-VSM capable
partition (i.e., in VTL0) and to use nested virtualization to run the
target guest OS.

To achieve this, we build the existing `virt_kvm` backend into the
paravisor openvmm build. So this has all the same limitations as
`virt_kvm` has on the host.

With this, it becomes possible to do some basic OpenHCL testing on a
Linux platform, with no Windows dependencies. Of course, since we are
using nested virtualization, there are some limitations:

* The host is not aware of the L2 guest, so it cannot offer vmbus or
  assigned devices directly to it. All devices must be emulated in the
  paravisor. There is no support for an interactive graphics console or
  anything else we haven't built a relay for yet.

* Performance is pretty bad in my testing. Of course, I tested this with
  3x nested virtualization, so that is not too surprising. I expect this
  to be much better on a native Linux host.

Currently, we still depend on the GET and other Hyper-V HCL devices, so
this can really only be used in conjunction with openvmm on the host.
Future changes can support alternate configuration mechanisms so that
this would be able to run in qemu or another VMM.

Also, virt_whp and virt_hvf do not yet support nested virtualization
(even though WHP and HVF do), so even with openvmm, this only works on
Linux.

Finally, I only built and tested this on x64, because I do not have a
KVM ARM64 environment that supports nested virtualization.